# python: updated versions + enable default optimizations