--- a/var/spack/repos/builtin/packages/python/package.py
+++ b/var/spack/repos/builtin/packages/python/package.py
@@ -145,14 +145,17 @@ class Python(AutotoolsPackage):
         dep_pfxs = [dspec.prefix for dspec in spec.dependencies('link')]
         config_args = [
             '--with-threads',
-            '--enable-optimizations',
             'CPPFLAGS=-I{0}'.format(' -I'.join(dp.include for dp in dep_pfxs)),
             'LDFLAGS=-L{0}'.format(' -L'.join(dp.lib for dp in dep_pfxs)),
         ]
+
+        if '@2.7.13:2.8,3.5.3:' in spec:
+            config_args.append('--enable-optimizations')
+
         if spec.satisfies('%gcc platform=darwin'):
             config_args.append('--disable-toolbox-glue')
 
-        if '%intel' in spec:
+        if '%intel' in spec and '@2.7.12:2.8,3.5.2:' in spec:
             config_args.append('--with-icc')
 
         if '+shared' in spec: