Enable default optimizations for production builds. Uses `--with-icc` if built with Intel compilers.