@xjrc 2018-03-15
After investigating these new flags a bit, I found that they're constrained to the following version ranges:

- `--with-icc`: @2.7.12:2.8,3.5.2:
- `--enable--optimizations`: @2.7.13:2.8,3.5.3:

It looks like the installs will work for versions where these flags are missing (e.g. for 2.7.10 a warning is printed but the build continues), but I think it's best for us to be explicit in the Spack script on when these flags can be applied. My primary concern is build functionality for older versions, so I'm not terribly concerned either way as these flags don't impact older versions; just a point worth that I think is worth considering.