@lee218llnl 2018-03-15
Do we really need to add --with-icc for Intel? In the configure script, it looks like this just sets CC and CXX to icc and icpc.

The "--enable-optimizations" flag looks to be pretty expensive for the build time of Python. Do we want to have this enabled by default? Perhaps we could also make it a variant so that folks can turn it off if they just want a quick build of Python.

If you do want either of these flags, please do add the version checks that @xjrc mentioned.