@alalazo 2018-03-15
@lee218llnl I have no strong opinion on making `--enable-optimizations` a variant. In case, though I think it should be enabled by default as it is reported to give a speed-up around ~10% (see the benchmark [here](https://www.activestate.com/blog/2014/06/python-performance-boost-using-profile-guided-optimization) where they apply the same optimization manually). Let me know your preference and I'll implement it.

For `--with-icc` I think it's worth it - we incur no penalties and if, in the future, the logic will start to do more things than just setting `CC` and `CXX` we'll be already up to date.