@adamjstewart 2018-03-15
~10% speedup makes it worth it in my mind. Spack is designed for supercomputers, so optimization is important. Definitely make it a variant though, a lot of people only need it as a build dependency. I would default to True personally.

@lee218llnl how bad is the build time increase?