@lee218llnl 2018-03-15
@alalazo I'm not entirely opposed to the "--with-icc" flag, but it does cause a difference in my environment. Wit that flag Python doesn't build explicitly with the full path to Spack's env/intel/icc:

==> Executing phase: 'build'
==> 'make' '-j72'
icc -pthread -c .......

and ultimately I see this in the resulting installed interpreter:

>>> distutils.sysconfig.get_config_vars()['CC']
'/usr/tce/bin/icc-18.0.0 -pthread'

And without the "--with-icc" flag I see:

==> Executing phase: 'build'
==> 'make' '-j72'
/nfs/tmp2/lee218/spack/lib/spack/env/intel/icc -pthread -c .......

And:

>>> distutils.sysconfig.get_config_vars()['CC']
'icc -pthread'

Granted Spack's env directory should be in the environment when building things under Spack, so it should not make a difference there, but is it really worth overriding Spack's CC and CXX commands?

@adamjstewart, I'm already over 2 hours and a test build is still going. I will update you when it's done, but no I am inclined to say that I don't think "--enable-optimizations" should be the default.