@alalazo 2018-03-16
On my laptop:
```console
$ uname -a
Linux nuvolari 4.4.0-116-generic #140~14.04.1-Ubuntu SMP Fri Feb 16 09:25:20 UTC 2018 x86_64 x86_64 x86_64 GNU/Linux

$ gcc --version
gcc (Ubuntu 4.8.4-2ubuntu1~14.04.4) 4.8.4
Copyright (C) 2013 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```
with `--enable-optimizations`:
```console
$ spack install -v python
...
==> Successfully installed python
  Fetch: 0.04s.  Build: 4m 49.40s.  Total: 4m 49.44s.
[+] /home/mculpo/PycharmProjects/spack/opt/spack/linux-ubuntu14.04-x86_64/gcc-4.8/python-2.7.14-qyytendyiuypwizustouwidjta4nhoce
```
without the option:
```console
$ spack install -v python
...
==> Successfully installed python
  Fetch: 0.04s.  Build: 2m 2.83s.  Total: 2m 2.87s.
[+] /home/mculpo/PycharmProjects/spack/opt/spack/linux-ubuntu14.04-x86_64/gcc-4.8/python-2.7.14-qyytendyiuypwizustouwidjta4nhoce
```
I'll check what is the timing difference on the clusters and post here.