@alalazo 2018-03-16
@lee218llnl Sorry if the question is silly, but from what you posted above you are compiling with `make -j 72` and Intel. Was wondering if the 2 hours are coming from a "competition" over the compiler license.