@alalazo 2018-03-16
On our clusters:
```console
$ uname -a
Linux fidis 3.10.0-514.36.5.el7.x86_64 #1 SMP Thu Dec 28 21:42:18 EST 2017 x86_64 x86_64 x86_64 GNU/Linux

$ which gcc
~/installation-tests/candidate/spack/opt/spack/linux-rhel7-x86_E5v4_Mellanox/gcc-4.8.5/gcc-7.3.0-ptts5bpzz3sqtkxvou5wb3ur7gu62e6i/bin/gcc

$ gcc --version
gcc (GCC) 7.3.0
Copyright (C) 2017 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```
with `--enable-optimizations`:
```console
$ spack install -v -j 4 python
... 
==> Successfully installed python
  Fetch: 0.12s.  Build: 5m 14.16s.  Total: 5m 14.29s.
[+] /home/culpo/installation-tests/candidate/spack/opt/spack/linux-rhel7-x86_E5v4_Mellanox/gcc-7.3.0/python-2.7.14-podgf6mwifnvsamjbuxlexmthxgzdr5l
```
and without:
```
$ spack install -v -j 4 python
...
==> Successfully installed python
  Fetch: 0.04s.  Build: 3m 1.35s.  Total: 3m 1.39s.
[+] /home/culpo/installation-tests/candidate/spack/opt/spack/linux-rhel7-x86_E5v4_Mellanox/gcc-7.3.0/python-2.7.14-podgf6mwifnvsamjbuxlexmthxgzdr5l
```
The option seems, more or less, to double the compilation time but I don't see anything near hours of compilation. Will try `python@3.X` next.