@alalazo 2018-03-16
@adamjstewart @lee218llnl `python@3.6.4` behaves the same as `python@2.7.14` as far as compilation time is concerned. Will try again with Intel on our clusters but, if you can confirm the results above, I would be for not even having a variant. 5 minutes of compilation for an install are imho completely fine.