@alalazo 2018-03-16
Final result on our cluster, with `--enable-optimizations` + `%intel`:
```console
$ spack install  -j 4 -v python%intel
...
==> Successfully installed python
  Fetch: 0.13s.  Build: 5m 16.36s.  Total: 5m 16.49s.
[+] /home/culpo/installation-tests/candidate/spack/opt/spack/linux-rhel7-x86_E5v4_Mellanox/intel-17.0.2/python-2.7.14-uefvmc4qgwztqzkfqs3tst57yocruqzu
```
For `--with-icc`: I tried to build `py-numpy` on top of this python, and the compilation went fine (or at least, there was nothing obviously wrong).