@alalazo 2018-03-16
I think I am finished with updating python from my side. Let me know if there are requests on variants to be added or removed and their defaults.