@adamjstewart 2018-03-16
I'm fine with doubling install times by default, but would still like to see a variant that explains that the result runs faster but takes longer to build. I'm still curious about what is going on with @lee218llnl's crazy build times though.