@xjrc 2018-03-16
Normally I'd not be too pleased with a ~1.6 times slowdown for builds by default, but given the normal build time is so low at ~3 minutes and the option grants a ~10-20%  performance improvement, it seems fine to me. Like @adamjstewart, I'd like to see it be a variant though; perhaps it can act as the starting point for a `+debug` variant.